fix model of price - datatype

+++ Services/Models/PriceModel.cs
@@ -22,13 +22,13 @@ namespace Services.Models
 
         public int id { get; set; }
         public string uid { get; set; }
-        public int number { get; set; }
+        public double number { get; set; }
         public string leading_zero_number { get; set; }
-        public float decimal_ { get; set; }
-        public float normal { get; set; }
+        public double decimal_ { get; set; }
+        public double normal { get; set; }
         public string hexadecimal { get; set; }
-        public float positive { get; set; }
-        public float negative { get; set; }
+        public double positive { get; set; }
+        public double negative { get; set; }
         public int non_zero_number { get; set; }
         public int digit { get; set; } 
 }
